+++ test/neat-client-test.el
@@ -132,20 +132,22 @@
           (expect (neat-bencode-get decoded "code") :to-equal "(+ 1 2)")
           (expect (neat-bencode-get decoded "session") :to-equal "S-1")))))
 
-  (it "includes file/line/column when provided"
+  (it "includes file/line/column/ns when provided"
     (let ((conn (neat-connection--make))
           sent)
       (cl-letf (((symbol-function 'process-live-p) (lambda (_) t))
                 ((symbol-function 'process-send-string)
                  (lambda (_p s) (setq sent s))))
-        (neat-eval conn "(+ 1 2)" nil "/tmp/foo.clj" 42 7)
+        (neat-eval conn "(+ 1 2)"
+                   :file "/tmp/foo.clj" :line 42 :column 7 :ns "my.ns")
         (let ((decoded (car (neat-bencode-decode sent))))
           (expect (neat-bencode-get decoded "file")
                   :to-equal "/tmp/foo.clj")
           (expect (neat-bencode-get decoded "line") :to-equal 42)
-          (expect (neat-bencode-get decoded "column") :to-equal 7)))))
+          (expect (neat-bencode-get decoded "column") :to-equal 7)
+          (expect (neat-bencode-get decoded "ns") :to-equal "my.ns")))))
 
-  (it "omits file/line/column when not provided"
+  (it "omits file/line/column/ns when not provided"
     (let ((conn (neat-connection--make))
           sent)
       (cl-letf (((symbol-function 'process-live-p) (lambda (_) t))
@@ -155,7 +157,8 @@
         (let ((decoded (car (neat-bencode-decode sent))))
           (expect (assoc "file" decoded) :to-be nil)
           (expect (assoc "line" decoded) :to-be nil)
-          (expect (assoc "column" decoded) :to-be nil))))))
+          (expect (assoc "column" decoded) :to-be nil)
+          (expect (assoc "ns" decoded) :to-be nil))))))
 
 (describe "neat-load-file"
   (it "builds a load-file op with contents and metadata"
@@ -165,7 +168,8 @@
       (cl-letf (((symbol-function 'process-live-p) (lambda (_) t))
                 ((symbol-function 'process-send-string)
                  (lambda (_p s) (setq sent s))))
-        (neat-load-file conn "(ns foo) (def x 1)" "/tmp/foo.clj" "foo.clj")
+        (neat-load-file conn "(ns foo) (def x 1)"
+                        :file-path "/tmp/foo.clj" :file-name "foo.clj")
         (let ((decoded (car (neat-bencode-decode sent))))
           (expect (neat-bencode-get decoded "op") :to-equal "load-file")
           (expect (neat-bencode-get decoded "file")